requirements.txt
dhanhq==2.0.2
python-telegram-bot==13.15
requests==2.32.3
pandas==2.2.3
python-dotenv==1.0.1
